split up the strict and loose versions of async callback
